di-demo-component-scan/src/main/java/com/dev/controllers/PropertyInjectedController.java
package com.dev.controllers;

import com.services.GreetingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Controller;

/**
 * Created by Jakez on 23/02/2018.
 */

// This is a case study of "NOT TO BE DONE" in this way. This is bad practice
// Injecting implementation is bad

@Controller
public class PropertyInjectedController {

    @Autowired
    @Qualifier("greetingServiceImpl")
    public GreetingService greetingServiceImpl;

    public String sayHello() {
        return greetingServiceImpl.sayGreeting();
    }

}
