di-demo-component-scan/src/test/java/com/dev/controllers/PropertyInjectedControllerTest.java
package com.dev.controllers;

import com.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

/**
 * Created by Jakez on 23/02/2018.
 */
public class PropertyInjectedControllerTest {

    private PropertyInjectedController propertyInjectedController;

    @Before
    public void setUp() throws Exception{
        this.propertyInjectedController = new PropertyInjectedController();
        this.propertyInjectedController.greetingServiceImpl = new GreetingServiceImpl();
    }

    @Test
    public void greeting() {
        assertEquals(GreetingServiceImpl.HELLO, propertyInjectedController.sayHello());
    }

}
